Only cube forward/back motion, add arm pos diagnostics
The mecanum wheels exhibited greater lag when rotating against each other (strafing and rotating) but not forward driving, so the gamepad cubing function was moved to only affect the drive variable.  Also added print statements for the position of the arm motors so we may begin work on implementing discrete arm positions.